backend/requirements.txt
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
asyncpg==0.30.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
fastapi==0.115.12
filelock==3.18.0
fonttools==4.58.2
fsspec==2025.5.1
ftfy==6.3.1
greenlet==3.2.3
grpcio==1.73.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
minio==7.2.15
mpmath==1.3.0
networkx==3.5
numpy==2.3.0
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
pillow==11.2.1
portalocker==2.10.1
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
qdrant-client
regex==2024.11.6
requests==2.32.4
scikit-learn==1.5.2
scipy==1.15.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.3.153
ultralytics-thop==2.0.14
urllib3==1.26.20
uvicorn==0.34.3
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
pytest-asyncio==1.0.0